Added more types of accessories
