--- Assets/Scripts/AccessoryEvents/UIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c0205ab267143b04abf89acf2c4a10b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
